jupyterBridge/server/test_ws.py
import websocket
import json
import time
import threading

def on_message(ws, message):
    print(f"Received: {message}")

def on_error(ws, error):
    print(f"Error: {error}")

def on_close(ws, close_status_code, close_msg):
    print(f"Closed: {close_status_code} {close_msg}")

def on_open(ws):
    print("Opened connection")
    # Send echo
    ws.send(json.dumps({"type": "input", "data": "echo Hello\n"}))

if __name__ == "__main__":
    websocket.enableTrace(True)
    ws = websocket.WebSocketApp("ws://localhost:8000/ws/terminal",
                              on_open=on_open,
                              on_message=on_message,
                              on_error=on_error,
                              on_close=on_close)

    ws.run_forever()
