jupyterBridge/server/requirements.txt
duckdb
pandas
pyarrow
rich
fastapi
uvicorn
python-multipart
websockets
dotenv
requests
jinja2
jupyterlab